pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "btc-exchange-intel-agent"
version = "0.1.0"
description = "BTC exchange address discovery agent with collector jobs and query API"
requires-python = ">=3.11"
dependencies = [
  "beautifulsoup4>=4.12,<5",
  "curl-cffi>=0.14,<0.15",
  "fastapi>=0.115,<1",
  "httpx>=0.27,<0.28",
  "lxml>=5.2,<6",
  "pydantic>=2.8,<3",
  "python-dotenv>=1.0,<2",
  "PyYAML>=6.0,<7",
  "sqlalchemy>=2.0,<3",
  "uvicorn[standard]>=0.30,<1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
